Update Pytorch support
1. Moving print to array from keras to global hls writer
2. Support pytorch wiht cpu or gpu model
